build: injects assets and color log output
